Revert "Fixed?"
This reverts commit d0ea2ae4858d3ce017e1de92e65031cd6e8b0315.

+++ Amphibian/Systems/Rendering/Sprites/Elements.cs
@@ -34,8 +34,8 @@ namespace Amphibian.Systems.Rendering.Sprites
         int Y { get; set; }
         int Height { get; set; }
         int Width { get; set; }
-        int OriginX { get; set; }
-        int OriginY { get; set; }
+        float OriginX { get; set; }
+        float OriginY { get; set; }
     }
 
     public interface ISpriteListElement
